Added JSON endpoint to get a redirect, added documentation, and other minor changes
Added the GET /{id}/json endpoint to make it easier to attain a redirect for anyone using the API.

Fully documented the API in README.md to make using it significantly easier.

Added error field to all JSON responses, which is included on all internal server error responses.

Changed JSON response values to lowercase to follow convention (I'm unsure why I initially opted for uppercase).

Removed hardcoded hostname values (qshr.tn) in the web client and in the server. Opting instead to use an environment variable (DOMAIN) on the server.

Slightly refactored the /new/ endpoint to make it simpler.