Added more documentation, simplified Docker Compose, and minor changes
Added an index to the README.md, also added documentation on how to host the site yourself.

Added a simplified Docker Compose which also includes MongoDB to simplify installation and upkeep.

Removed DOMAIN environment variable, instead opting to get the hostname from a request.

Changed JS to not use the name qshr.tn, but instead assume it from the hostname.